refactor: #513 recyclerview adapter와 ViewHolder 분리
