+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/view/discussions/hot/viewholder/HotPopularDiscussionViewHolder.kt
@@ -1,4 +1,4 @@
-package com.team.todoktodok.presentation.view.discussions.hot.adapter
+package com.team.todoktodok.presentation.view.discussions.hot.viewholder
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,13 +6,15 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.team.todoktodok.databinding.ItemHotPopularDiscussionBinding
 import com.team.todoktodok.presentation.view.discussions.adapter.DiscussionAdapter
+import com.team.todoktodok.presentation.view.discussions.hot.adapter.HotDiscussionItems
 
 class HotPopularDiscussionViewHolder private constructor(
     binding: ItemHotPopularDiscussionBinding,
     handler: Handler,
 ) : RecyclerView.ViewHolder(binding.root) {
     private val discussionAdapter = DiscussionAdapter(handler)
-    private val manager = LinearLayoutManager(itemView.context, LinearLayoutManager.HORIZONTAL, false)
+    private val manager =
+        LinearLayoutManager(itemView.context, LinearLayoutManager.HORIZONTAL, false)
 
     init {
         with(binding) {
